change splashscreen to functional component
