+++ src/pages/SplashScreen/index.js
@@ -1,67 +1,57 @@
-import { Text, StyleSheet, View, Image, BackHandler } from "react-native";
-import React, { Component } from "react";
+import { Text, StyleSheet, View, Image } from "react-native";
+import React, { useEffect } from "react";
 
 import Logo from "../../assets/img/Logo&Name.png";
+import { useState } from "react";
 
-export default class SplashScreen extends Component {
-  state = {
-    status: 0,
-    widthLoading: 0,
-  };
+const SplashScreen = ({ navigation }) => {
+  const [isLoad, setIsLoad] = useState(false);
+  const [status, setStatus] = useState(0);
+  const [widthLoading, setWidthLoading] = useState(0);
 
-  componentDidMount() {
-    // console.log(this.props);
-    BackHandler.addEventListener("hardwareBackPress", this.handleBackButton);
+  useEffect(() => {
+    if (isLoad === false) {
+      setIsLoad(true);
+      statusLoading();
+    }
+  });
 
-    var i = 1;
+  const statusLoading = () => {
+    var i = 0; //  set your counter to 1
+    var wid = 0;
 
-    const goToLogin = () => {
-      this.props.navigation.navigate("Login");
-    };
-
-    const myLoop = () => {
-      this.setState({
-        status: this.state.status + 1,
-        widthLoading: this.state.widthLoading + 2,
-      });
-      // console.log("State Log", this.state.widthLoading);
+    function myLoop() {
+      setStatus(i);
+      setWidthLoading(wid);
       setTimeout(function () {
-        // console.log("hello", i);
+        // console.log("hello", wid);
+        wid = wid + 2;
         i++;
         if (i < 101) {
           myLoop();
         } else {
-          goToLogin();
-          // console.log("Navigate");
+          navigation.navigate("Login");
         }
       }, 0);
-    };
+    }
 
     myLoop();
-  }
+  };
 
-  handleBackButton() {
-    // ToastAndroid.show("Back button is pressed", ToastAndroid.SHORT);
-    alert("Can't");
-    return true;
-  }
-
-  render() {
-    return (
-      <View style={styles.container}>
-        <View style={{ alignItems: "center" }}>
-          <Image source={Logo} style={styles.logoKedasi} />
-          <Text style={{ marginTop: 50 }}>Loading {this.state.status}%</Text>
-          <View style={{ width: 200 }}>
-            <View
-              style={[styles.loading, { width: this.state.widthLoading }]}
-            />
-          </View>
+  return (
+    <View style={styles.container}>
+      <View style={{ alignItems: "center" }}>
+        <Image source={Logo} style={styles.logoKedasi} />
+        <Text style={{ marginTop: 50 }}>Loading {status}%</Text>
+        <View style={{ width: 200 }}>
+          <View style={[styles.loading, { width: widthLoading }]} />
         </View>
       </View>
-    );
-  }
-}
+    </View>
+  );
+};
+
+export default SplashScreen;
 
 const styles = StyleSheet.create({
   container: {
